+++ src/setup.jl
@@ -76,5 +76,6 @@ export subset_amsr
 export T90_from_T48
 export T90_from_T68
 export toc
+export xyz_to_enu
 export z_from_pressure